Added special case for IRQ firing after hw_halt() / terminate_all().
After terminate_all() there is no active transaction anymore, however the IRQ can still fire once!

This is caused by the way terminate_all() is implemented and the Xilinx DMA soft IP core's halt mechanism.  Terminate all tells the HW core to halt, and frees all of the descriptors (active, pending and completed).  But the HW core will not halt until after the active transaction is completed, and since we cannot guarantee there will be one, we cannot wait for this to occur.

There is a potential danger here:
If the client calls terminate_all() and then frees the DMA buffer backing the active transaction, the HW core will still try to complete the active transaction. In order to prevent this, the DMA core needs to be reset in terminate_all() but that resets BOTH the channels! FUBAR.

Fortunately, my axis-reader module does not free DMA buffers unless it is being unloaded.